refactor: moved getMatcheeConversation from helper to conversation.ts
