+++ graphql/match/mutations.ts
@@ -6,8 +6,7 @@ import { dissolveMatch, reactivateMatch } from '../../common/match/dissolve';
 import { createMatch } from '../../common/match/create';
 import { GraphQLContext } from '../context';
 import { ConcreteMatchPool, pools } from '../../common/match/pool';
-import { getMatcheeConversation } from '../../common/chat/helper';
-import { markConversationAsWriteable } from '../../common/chat';
+import { getMatcheeConversation, markConversationAsWriteable } from '../../common/chat';
 import { JSONResolver } from 'graphql-scalars';
 import { createAdHocMeeting } from '../../common/appointment/create';
 import { AuthenticationError } from '../error';
